+++ apps/one_piece_commanded/lib/one_piece/commanded/aggregate/stateless_lifespan.ex
@@ -5,7 +5,24 @@ defmodule OnePiece.Commanded.Aggregate.StatelessLifespan do
 
   @behaviour Commanded.Aggregates.AggregateLifespan
 
+  @doc """
+  Stops the aggregate after a command.
+      iex> OnePiece.Commanded.Aggregate.StatelessLifespan.after_command(%MyCommandOne{})
+      :stop
+  """
   def after_command(_command), do: :stop
+
+  @doc """
+  Stops the aggregate after an event.
+      iex> OnePiece.Commanded.Aggregate.StatelessLifespan.after_event(%DepositAccountOpened{})
+      :stop
+  """
   def after_event(_event), do: :stop
+
+  @doc """
+  Stops the aggregate after an error.
+      iex> OnePiece.Commanded.Aggregate.StatelessLifespan.after_error({:error, :something_happened})
+      :stop
+  """
   def after_error(_error), do: :stop
 end