+++ apps/one_piece_commanded/test/one_piece/commanded/command_test.exs
@@ -7,4 +7,19 @@ defmodule OnePiece.Commanded.CommandTest do
       assert %{uuid: ["can't be blank"]} = TestSupport.errors_on(changeset)
     end
   end
+
+  test "allow custom type as the identity" do
+    uuid =
+      TestSupport.AccountNumber.new!(%{
+        account_number: "123e4567-e89b-12d3-a456-426655440000",
+        branch: "123e4567-e89b-12d3-a456-426655440000"
+      })
+
+    assert {:ok, _command} = TestSupport.OpenDepositAccountCommand.new(%{uuid: uuid, type: :DEPOSITORY})
+  end
+
+  test "allow custom type as the identity fails when invalid value is passed" do
+    {:error, changeset} = TestSupport.OpenDepositAccountCommand.new(%{uuid: "123e4567-e89b-12d3-a456-426655440000"})
+    assert %{uuid: ["is invalid"]} = TestSupport.errors_on(changeset)
+  end
 end